server/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
psycopg[asyncio]
python-dotenv
redis
google-genai
pydantic[email]
alembic
asyncpg
